move processor into an internal package, in preparation for profiles
