+++ processor/internal/factory.go
@@ -1,0 +1,172 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal // import "go.opentelemetry.io/collector/processor/internal"
+
+import (
+	"context"
+
+	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/consumer"
+)
+
+// Internal is a Factory interface for processors.
+//
+// This interface cannot be directly implemented. Implementations must
+// use the NewProcessorFactory to implement it.
+type Internal interface {
+	component.Factory
+
+	// CreateTracesProcessor creates a TracesProcessor based on this config.
+	// If the processor type does not support tracing or if the config is not valid,
+	// an error will be returned instead.
+	CreateTracesProcessor(ctx context.Context, set Settings, cfg component.Config, nextConsumer consumer.Traces) (Traces, error)
+
+	// TracesProcessorStability gets the stability level of the TracesProcessor.
+	TracesProcessorStability() component.StabilityLevel
+
+	// CreateMetricsProcessor creates a MetricsProcessor based on this config.
+	// If the processor type does not support metrics or if the config is not valid,
+	// an error will be returned instead.
+	CreateMetricsProcessor(ctx context.Context, set Settings, cfg component.Config, nextConsumer consumer.Metrics) (Metrics, error)
+
+	// MetricsProcessorStability gets the stability level of the MetricsProcessor.
+	MetricsProcessorStability() component.StabilityLevel
+
+	// CreateLogsProcessor creates a LogsProcessor based on the config.
+	// If the processor type does not support logs or if the config is not valid,
+	// an error will be returned instead.
+	CreateLogsProcessor(ctx context.Context, set Settings, cfg component.Config, nextConsumer consumer.Logs) (Logs, error)
+
+	// LogsProcessorStability gets the stability level of the LogsProcessor.
+	LogsProcessorStability() component.StabilityLevel
+
+	unexportedFactoryFunc()
+}
+
+// FactoryOption apply changes to Options.
+type FactoryOption interface {
+	// applyProcessorFactoryOption applies the option.
+	applyProcessorFactoryOption(o *factory)
+}
+
+var _ FactoryOption = (*factoryOptionFunc)(nil)
+
+// factoryOptionFunc is a FactoryOption created through a function.
+type factoryOptionFunc func(*factory)
+
+func (f factoryOptionFunc) applyProcessorFactoryOption(o *factory) {
+	f(o)
+}
+
+// CreateTracesFunc is the equivalent of Factory.CreateTraces().
+type CreateTracesFunc func(context.Context, Settings, component.Config, consumer.Traces) (Traces, error)
+
+// CreateTracesProcessor implements Factory.CreateTracesProcessor().
+func (f CreateTracesFunc) CreateTracesProcessor(
+	ctx context.Context,
+	set Settings,
+	cfg component.Config,
+	nextConsumer consumer.Traces) (Traces, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
+
+// CreateMetricsFunc is the equivalent of Factory.CreateMetrics().
+type CreateMetricsFunc func(context.Context, Settings, component.Config, consumer.Metrics) (Metrics, error)
+
+// CreateMetricsProcessor implements Factory.CreateMetricsProcessor().
+func (f CreateMetricsFunc) CreateMetricsProcessor(
+	ctx context.Context,
+	set Settings,
+	cfg component.Config,
+	nextConsumer consumer.Metrics,
+) (Metrics, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
+
+// CreateLogsFunc is the equivalent of Factory.CreateLogs().
+type CreateLogsFunc func(context.Context, Settings, component.Config, consumer.Logs) (Logs, error)
+
+// CreateLogsProcessor implements Factory.CreateLogsProcessor().
+func (f CreateLogsFunc) CreateLogsProcessor(
+	ctx context.Context,
+	set Settings,
+	cfg component.Config,
+	nextConsumer consumer.Logs,
+) (Logs, error) {
+	if f == nil {
+		return nil, component.ErrDataTypeIsNotSupported
+	}
+	return f(ctx, set, cfg, nextConsumer)
+}
+
+type factory struct {
+	cfgType component.Type
+	component.CreateDefaultConfigFunc
+	CreateTracesFunc
+	tracesStabilityLevel component.StabilityLevel
+	CreateMetricsFunc
+	metricsStabilityLevel component.StabilityLevel
+	CreateLogsFunc
+	logsStabilityLevel component.StabilityLevel
+}
+
+func (f *factory) Type() component.Type {
+	return f.cfgType
+}
+
+func (f *factory) unexportedFactoryFunc() {}
+
+func (f factory) TracesProcessorStability() component.StabilityLevel {
+	return f.tracesStabilityLevel
+}
+
+func (f factory) MetricsProcessorStability() component.StabilityLevel {
+	return f.metricsStabilityLevel
+}
+
+func (f factory) LogsProcessorStability() component.StabilityLevel {
+	return f.logsStabilityLevel
+}
+
+// WithTraces overrides the default "error not supported" implementation for CreateTraces and the default "undefined" stability level.
+func WithTraces(createTraces CreateTracesFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factory) {
+		o.tracesStabilityLevel = sl
+		o.CreateTracesFunc = createTraces
+	})
+}
+
+// WithMetrics overrides the default "error not supported" implementation for CreateMetrics and the default "undefined" stability level.
+func WithMetrics(createMetrics CreateMetricsFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factory) {
+		o.metricsStabilityLevel = sl
+		o.CreateMetricsFunc = createMetrics
+	})
+}
+
+// WithLogs overrides the default "error not supported" implementation for CreateLogs and the default "undefined" stability level.
+func WithLogs(createLogs CreateLogsFunc, sl component.StabilityLevel) FactoryOption {
+	return factoryOptionFunc(func(o *factory) {
+		o.logsStabilityLevel = sl
+		o.CreateLogsFunc = createLogs
+	})
+}
+
+// NewFactory returns a Factory.
+func NewFactory(cfgType component.Type, createDefaultConfig component.CreateDefaultConfigFunc, options ...FactoryOption) Internal {
+	f := &factory{
+		cfgType:                 cfgType,
+		CreateDefaultConfigFunc: createDefaultConfig,
+	}
+	for _, opt := range options {
+		opt.applyProcessorFactoryOption(f)
+	}
+	return f
+}
